feat(theme-builder): move colors.ts into a core/ subdirectory
